cr6-groupe4/requirements.txt
numpy < 1.24
networkx
pydot
matplotlib
pycryptodome
argparse